add function to get first person interface from weapon
move model node struct
