+++ xlive/H2MOD/Variants/GraveRobber/GraveRobber.cpp
@@ -214,7 +214,7 @@ bool GraveRobber::c_game_statborg__adjust_player_stat(ExecTime execTime, c_game_
 
 bool GraveRobber::OnAutoPickupHandler(ExecTime execTime, datum playerIdx, datum objectIdx)
 {
-	const s_weapon_data_definition* weaponObject = object_get_fast_unsafe<s_weapon_data_definition>(objectIdx);
+	const weapon_datum* weapon = object_get_fast_unsafe<weapon_datum>(objectIdx);
 
 	bool handled = false;
 
@@ -222,7 +222,7 @@ bool GraveRobber::OnAutoPickupHandler(ExecTime execTime, datum playerIdx, datum 
 	{
 	case ExecTime::_preEventExec:
 
-		if (DATUM_INDEX_TO_ABSOLUTE_INDEX(weaponObject->item.object.tag_definition_index) == DATUM_INDEX_TO_ABSOLUTE_INDEX(e_weapons_datum_index::ball))
+		if (DATUM_INDEX_TO_ABSOLUTE_INDEX(weapon->item.object.tag_definition_index) == DATUM_INDEX_TO_ABSOLUTE_INDEX(e_weapons_datum_index::ball))
 		{
 			GraveRobber::PickupSkull(playerIdx, objectIdx);
 			handled = true;
